pyproject.toml
[project]
name = "simple-experiment-manager"
version = "0.1.0"
authors = [
    { name = "toshitaka uchino", email = "toshitaka.uchino@gmail.com" }
]
license = { text = "MIT" }
description = "Experiment label management tool based on Pydantic"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.12.5",
    "ruamel-yaml>=0.18.17",
    "typer>=0.12.0",
    "rich>=13.0.0"
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
]
